fix: Resolve clippy warnings in tests and examples
- Replace std::sync::Arc::from with StringValue::from for SqlValue::Varchar
- Add StringValue import to affected files
- Use const initializer for thread_local values
- Replace assert!(true, ...) with empty match arms
- Use .is_multiple_of() instead of % n == 0
- Use .div_ceil() instead of manual ceiling division
- Simplify nested format! calls
- Change assert_eq!(x, true) to assert!(x)
- Use !.is_empty() instead of .len() > 0
- Use if let Ok(x) instead of if let Some(x) = .ok()
- Use .contains() instead of .iter().any()
- Use .as_str() instead of ambiguous .as_ref() for ArcStr
- Remove unnecessary as f64 casts
- Use .first() instead of .get(0)

+++ tests/regression/issue_990.rs
@@ -83,7 +83,7 @@ fn test_issue_990_multiple_unary_plus() {
         vibesql_types::SqlValue::Integer(n) => assert_eq!(*n, 97, "Expected 97, got {}", n),
         vibesql_types::SqlValue::Numeric(n) => {
             // Check if it's close to 97 (allowing for floating point comparison)
-            assert!(((*n as f64) - 97.0).abs() < 0.001, "Expected 97, got {}", n);
+            assert!((*n - 97.0).abs() < 0.001, "Expected 97, got {}", n);
         }
         other => panic!("Expected Integer or Numeric, got {:?}", other),
     }
@@ -158,7 +158,7 @@ fn test_issue_990_simpler_case() {
     match &result[0].values[0] {
         vibesql_types::SqlValue::Integer(n) => assert_eq!(*n, 97, "Expected 97, got {}", n),
         vibesql_types::SqlValue::Numeric(n) => {
-            assert!(((*n as f64) - 97.0).abs() < 0.001, "Expected 97, got {}", n);
+            assert!((*n - 97.0).abs() < 0.001, "Expected 97, got {}", n);
         }
         other => panic!("Expected Integer or Numeric, got {:?}", other),
     }
